+++ src/libs/api/index.ts
@@ -1,2 +1,1 @@
 export * from './axiosInstance';
-export * from './reGetToken';
